fix test scheduling 2
